% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{as.charTime}
\alias{as.charTime}
\title{as.charTime}
\usage{
as.charTime(simTime)
}
\arguments{
\item{simTime}{}
}
\value{
charTime A char containing a time in the format HH:MM
}
\description{
as.charTime
}

