% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{as.simDate}
\alias{as.simDate}
\title{Converts a DateTime to a simDate. truncates the Time part. range 0-364}
\usage{
as.simDate(datetime)
}
\arguments{
\item{datetime}{}
}
\value{
t A simDate
}
\description{
Converts a DateTime to a simDate. truncates the Time part. range 0-364
}

