% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{as.simDateTime}
\alias{as.simDateTime}
\title{Converts a DateTime to a simDateTime}
\usage{
as.simDateTime(datetime)
}
\arguments{
\item{datetime}{}
}
\value{
t A simDateTime
}
\description{
Converts a DateTime to a simDateTime
}

