% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{as.simTime}
\alias{as.simTime}
\title{Converts a DateTime to a simTime. truncates the Date part. range 0-(24*60*60-1)}
\usage{
as.simTime(datetime)
}
\arguments{
\item{datetime}{}
}
\value{
t A simDateTime
}
\description{
Converts a DateTime to a simTime. truncates the Date part. range 0-(24*60*60-1)
}

