% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtimer.R
\name{sim_date}
\alias{sim_date}
\title{Return day of year from rel. time in seconds.}
\usage{
sim_date(sim_datetime)
}
\arguments{
\item{sim_datetime}{A sim_datetime}
}
\value{
the day of the period (> 0, in many cases also < 365)
}
\description{
Return day of year from rel. time in seconds.
}
