% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{simDate}
\alias{simDate}
\title{Date part of a simDateTime}
\usage{
simDate(t)
}
\arguments{
\item{t}{A simDateTime}
}
\value{
the day of year (0-364)
}
\description{
Date part of a simDateTime
}

