% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{simTime}
\alias{simTime}
\title{Time part of a simDateTime}
\usage{
simTime(t)
}
\arguments{
\item{t}{A simDateTime}
}
\value{
the actual time in seconds (0-(24*60*60-1))
}
\description{
Time part of a simDateTime
}

