% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTimeR.R
\name{simWeekday}
\alias{simWeekday}
\title{Weekday of a simDateTime}
\usage{
simWeekday(t)
}
\arguments{
\item{t}{A simDateTime}
}
\value{
the weekday in the format: "Mo, Di, Mi, Do, Fr, Sa, So"
}
\description{
Weekday of a simDateTime
}

