% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtimer.R
\name{datetime}
\alias{datetime}
\title{Converts a sim_datetime to a datetime.}
\usage{
datetime(sim_datetime)
}
\arguments{
\item{sim_datetime}{}
}
\value{
datetime A POSIXct date-time
}
\description{
Converts a sim_datetime to a datetime.
}
