% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtimer.R
\name{sim_date}
\alias{sim_date}
\title{Return day of year from rel. time in seconds.}
\usage{
sim_date(sim_datetime)
}
\arguments{
\item{sim_datetime}{A sim_datetime}
}
\value{
the day of the period (> 0, in many cases also < 365)
}
\description{
Return day of year from rel. time in seconds.
}
\examples{
origin_date <- lubridate::ymd_hms("2016-01-01 00:00:00")
sim_date(sim_datetime(lubridate::ymd_hms("2016-01-01 00:01:00"), origin_date))
# [1] 0
sim_date(sim_datetime(lubridate::ymd_hms("2016-01-02 00:01:00"), origin_date))
# [1] 1
}
