% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtimer.R
\name{sim_datetime}
\alias{sim_datetime}
\title{Converts a datetime to a sim_datetime.}
\usage{
sim_datetime(datetime)
}
\arguments{
\item{datetime}{A datetime}
}
\value{
sim_datetime A sim_datetime
}
\description{
Converts a datetime to a sim_datetime.
}
