% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtimer.R
\name{sim_time}
\alias{sim_time}
\title{Day-time part of a sim_datetime}
\usage{
sim_time(sim_datetime)
}
\arguments{
\item{sim_datetime}{A sim_datetime}
}
\value{
the actual day-time in seconds (0-(24*60*60-1))
}
\description{
Day-time part of a sim_datetime
}
\examples{
origin_date <- lubridate::ymd_hms("2016-01-01 00:00:00")
sim_time(sim_datetime(lubridate::ymd_hms("2016-01-01 00:01:00")))
# [1] 60
sim_time(sim_datetime(lubridate::ymd_hms("2016-01-02 00:01:00")))
# [1] 60
}
