% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtimer.R
\name{sim_wday}
\alias{sim_wday}
\title{Weekday of a sim_datetime}
\usage{
sim_wday(sim_datetime)
}
\arguments{
\item{sim_datetime}{A sim_datetime}
}
\value{
the weekday in the format: "Mo, Di, Mi, Do, Fr, Sa, So"
}
\description{
Weekday of a sim_datetime
}
